Add src files, update README and LICENSE
These files came from my "my_nif" proof-of-concept project and I have no
idea if they compile or run. This is the first step towards converting
from the proof-of-concept to a library that I can use as a dependency in
an Erlang project to output Erlang source code to a PNG file.

I'm not entirely sure how to comply with the freetype2 license. I've
included the suggested credit/disclaimer in both the render.c file
derived from the example at
https://www.freetype.org/freetype2/docs/tutorial/example1.c and I've
included the same in the README and the LICENSE file. Hopefully that is
enough. I'll include it here just in case. I have no intention of
passing this code off as my own.

Portions of this software are copyright © <year> The FreeType
Project (www.freetype.org).  All rights reserved.

This project relies on FreeType2 as well, but I don't know enough C to
know whether the FreeType2 source or binaries are packaged up or need to
be on the machine where this runs.