Pass target image into bitmap drawing function
I copied a draw_image function from the net:
https://www.freetype.org/freetype2/docs/tutorial/example1.c

If I write to the global image variable in the draw_bitmap function then
once I exit the function the image variable contains garbage.
However, if I pass in the array then the values stay correct.

Added lots of debug output but it's commented out for now.

I have a theory about the transformations by pen: I think that just
changes the values reported by the metrics; I don't think it changes
where the glyph is drawn in the face->glyph->slot (or whatever).