Get two chars rendered, shift to baseline
Got two characters working. The pen just seems like a convenience
feature that helps you keep track of where you need to put the character
in your own memory, but it stays in the same place in the face glyph
slot.

Shift letters down according to the height minus the horizontal y
bearing. The horizontal y bearing is how much of the letter is above the
baseline. height - bearing = amount to shift down below baseline.

When I don't |= the bitmap values with the target memory values I get
garbage and I'm not sure why.