Alloc mem once & clear, align chars
There doesn't seem to be a need to malloc the memory for the image and
then have Erlang allocate the memory again. I was allocating it once,
setting it all to zero (apparently there's a calloc call that does this)
and then it was getting all corrupted again. Turns out the
enif_alloc_binary call was--I think--re-allocating memory. If I didn't
clear out the memory then my image was filled with junk alpha values.

The face->glyph->bitmap_top value is the distance from the baseline of
the glyph to the top of the glyph. I want the descenders of letters to
hang down below so I need to shift the hanging letters down. To do this
I'm subtracting bitmap_top from the HEIGHT and then subtracting 25.
HEIGHT - bitmap_top will put the baseline at HEIGHT, i.e. the bottom
of the target image. Adding 25 moves the baseline up to make room for
the descenders. Note: this is all specific to the exact font and
dimensions that I'm using. Eventually I'll need to find a general way to
calculate this.

See https://en.wikipedia.org/wiki/Typeface_anatomy for definitions of
what baseline and descenders are. (I didn't know)