Find .so in ebin, find char in argv[0], nif_funcs
Tell render_char.erl to look in the ebin directory for render_nif.so,
not in the current directory, which is the top level directory.

I used to send the number of chars and then the list of chars to
render_char, so chars was in argv[1]; however, I've dropped the number
of chars so now the (single) char is in argv[0].

Update the array of ErlNifFunc to point to the correct function now that
I've renamed it.