Return "canvas" width
A char array is passed to render.c and only part of the char array is
used to render the letter into. Either we need to generate a binary
that is exactly the right size to render to or we need to return how
big the "canvas" binary is so that the caller can extract the rendered
letter.  Since at this point it's easier to return the width of the
"canvas" binary I'm doing that. I don't even know if there's a way to
pre-calculate the correct size binary to render the letter into. I could
create a second char array in render_nif.c but I'm more comfortable in
Erlang so I'm going to do it there. (I'd say "for now" but realistically
I'll probably never change it.)